% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADCPpointCloud_functions.r
\name{addPlane.z}
\alias{addPlane.z}
\title{addPlane.z}
\usage{
addPlane.z(data, plane)
}
\arguments{
\item{data}{a \code{data.Frame} containing the coordinates of the points to be fitted to the plane.}

\item{plane}{a \code{vector} created by the \code{points2Plane} function}
}
\value{
a \code{data.Frame} containing the original coordinates and one \code{plane.z} column with the plane elevation for each point.
}
\description{
This function adds elevations based on fitted plane
}
\author{
Claude Flener \email{claude.flener@utu.fi}
}
\seealso{
\code{\link{points2Plane}}
}

