% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{conveyance}
\alias{conveyance}
\title{Channel conveyance}
\usage{
conveyance(n, A, R, Cm)
}
\arguments{
\item{n}{Manning's roughness coefficient (dimensionless).}

\item{A}{Flow area [\eqn{L^2}].}

\item{R}{Hydraulic radius [\eqn{L}].}

\item{Cm}{Unit conversion coefficient for Manning's equation. For SI units, Cm = 1.}
}
\value{
The channel conveyance.
}
\description{
Calculate the channel conveyance.
}
\details{
Channel conveyance is routinely calculated in numerical solutions of steady, 
  gradually-varied and unsteady flows. This function is used extensively by internal 
  procedures and is made accessible to the user for convenience.
}
