% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydemos.r
\name{demo_shiny}
\alias{demo_shiny}
\title{Shiny Demonstrations}
\usage{
demo_shiny(ex)
}
\arguments{
\item{ex}{Example to run.}
}
\description{
Demonstrate package functionality via Shiny apps
}
\details{
Demonstrations available:
  \code{"gvf"} Gradually-varied flow.
}
\examples{
\dontrun{
# get list of available demos
demo_shiny()
# run the gradually-varied flow demo
demo_shiny("gvf")
}

}

