% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functional.R
\name{distribution_iterator}
\alias{distribution_iterator}
\title{Distribution iterator}
\usage{
distribution_iterator(fun, n = 1, ...)
}
\arguments{
\item{fun}{A distribution function}

\item{n}{Number of things to sample. I am so sorry.}

\item{...}{Additional arguments to \code{fun}}
}
\description{
Distribution iterator
}

