% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/file_iterator.R
\name{file_iterator}
\alias{file_iterator}
\title{File iterator}
\usage{
file_iterator(filename, n = 1L)
}
\arguments{
\item{filename}{Filename}

\item{n}{Number of lines to yield each time}
}
\description{
File iterator
}

