% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functional.R
\name{functional_iterator}
\alias{functional_iterator}
\title{Functional iterator}
\usage{
functional_iterator(fun, ...)
}
\arguments{
\item{fun}{A function}

\item{...}{Additional arguments to \code{fun}}
}
\description{
Functional iterator
}

