% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/iapply.R
\name{iapply}
\alias{iapply}
\title{Iterator apply}
\usage{
iapply(x, fun, ...)
}
\arguments{
\item{x}{an iterator object}

\item{fun}{the function to be applied to each element}

\item{...}{optional arguments to \code{fun}}
}
\description{
Iterator apply
}

