% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/iterator.R
\name{iterator}
\alias{iterator}
\title{Construct a general iterator}
\usage{
iterator(object, ...)
}
\arguments{
\item{object}{Any R object}

\item{...}{additional parameters}
}
\description{
Construct a general iterator
}

