% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/random.R
\name{pick_iterator}
\alias{pick_iterator}
\title{Random sampling iterator}
\usage{
pick_iterator(it, pr)
}
\arguments{
\item{it}{Iterator}

\item{pr}{Probability of sampling each element}
}
\description{
Random sampling iterator
}

