% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADCPpointCloud_functions.r
\name{points2Plane}
\alias{points2Plane}
\title{points2Plane}
\usage{
points2Plane(points = NA)
}
\arguments{
\item{points}{a \code{data.Frame} containing the coordinates of three points in XYZ format.}
}
\value{
a \code{vector} containing the cross-product of three vectors and the offset d
}
\description{
This function finds the equation for computing a plane based on three 3D points
}
\details{
find the equation for fitting a plane: ax +by + cz = d
}
\author{
Claude Flener \email{claude.flener@utu.fi}
}

