% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADCPpointCloud_functions.r
\name{riv2FlowLayers}
\alias{riv2FlowLayers}
\title{riv2FlowLayers}
\usage{
riv2FlowLayers(path = getwd(), externalGPS = NULL, TimeShiftSecs = 0,
  plane = NULL, z.out = "Elevation", out = "all", outdir = ".",
  outfile = NULL, fileAppendix = "", verbose = TRUE, plot = TRUE,
  map = TRUE)
}
\arguments{
\item{path}{the path to the input files (default = working directory)}

\item{externalGPS}{\code{GNSS} if external GNSS data is to be used for
computing point locations (default = NULL)}

\item{TimeShiftSecs}{Optional: an integer as the number of seconds to shift GNSS data, if not in
UTC (default = 0)}

\item{plane}{Optional: a dataFrame output by \code{points2Plane()}. (default = NULL)}

\item{z.out}{Z coordinate of output file. Takes one of three strings as values: "Elevation", "plane.z", "bed.z" (default = "Elevation")}

\item{out}{Output filetype. can take either of these strings as values:
"csv", "shape", "all" (default), "none"}

\item{outdir}{target directory for output files (default = ".")}

\item{outfile}{Optional: a string as the file name base for output files}

\item{fileAppendix}{a string appendix to file name
fileAppendix is intended to be used
together with automatic file names (outfile=NULL) to allow
customisation of file names while retaining the convenience of
automatically generated file names}

\item{verbose}{logical, prints the first 100 lines of the result to screen
(default =TRUE)}

\item{plot}{logical, plots the points created. (default = TRUE)}

\item{map}{logical , creates PDF-maps of the bathymetry and flow data (default = TRUE)}
}
\value{
point data either as CSV, ESRI shape, or both (default)
automatically handles batch processing if path is a directory
}
\description{
This function creates point layers of flow direction and speed, and
bathymetry
Three flow layers: Surface, Depth-integrated (average), Bottom data
bathymetry: depth and bed elevation (and surface elevation for
post-processing)
}
\note{
If \code{plane} is provided, this adds theoretical

if out == "none", no output file is created but the spatialPointsDataFrame is returned on exit; all other options produce output files but return empty.

outfile defaults to NULL, meaning that the file name of the
output file is automatically created form the first ADCP *.mat file
used for input.
}
\author{
Claude Flener \email{claude.flener@utu.fi}
}

