% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADCPpointCloud_functions.r
\name{riv2FlowPointCloud}
\alias{riv2FlowPointCloud}
\title{riv2FlowPointCloud}
\usage{
riv2FlowPointCloud(path = getwd(), externalGPS = NULL, TimeShiftSecs = 0,
  plane = NULL, z.out = "Elevation", out = "all", outdir = ".",
  outfile = NULL, fileAppendix = "", verbose = TRUE)
}
\arguments{
\item{path}{the path to the input files (default = working directory.) Automatically handles batch processing if path is a directory.}

\item{externalGPS}{\code{GNSS} if external GNSS data is to be used for computing point locations (default = NULL)}

\item{TimeShiftSecs}{Optional: an integer as the number of seconds to shift GNSS data, if not in UTC (default = 0)}

\item{plane}{Optional: a dataFrame output by \code{points2Plane()}. (default = NULL)}

\item{z.out}{Z coordinate of output file. Takes one of three strings as values: "Elevation", "plane.z", "bed.z" (default = "Elevation")}

\item{out}{Output filetype. can take either of these strings as values: "csv", "shape", "all" (default), "none"}

\item{outdir}{target directory for output files (default = ".")}

\item{outfile}{Optional: a string as the file name base for output files}

\item{fileAppendix}{a string appendix to file name}

\item{verbose}{logical, prints the first 100 lines of the result to screen (default =TRUE)}
}
\value{
A \code{spatialPointsDataFrame}, or, a file containing 3D point data (XYZ) either as CSV, ESRI shape, or both (default). The flow and depth data, as well as a bunch of other variables from the ADCP are included to be used as attributes / scalar fields.
}
\description{
This function creates 3D point clouds of flow direction, speed, river bed and water surface based on RiverSurveyor ADCP data and, optionally, external GNSS data.
}
\note{
path: If you provide a full path to a RiverSurveyor *.mat file, that file will be used for processing and only the GNSS points form the external GNSS file (if applicable) will be used. If your measurement is split over several *.mat files, put them all in the same directory and input the path to the directory instead of a file. The function will batch-process all *.mat files in that directory and combine the resulting points into one output file.

If \code{plane} is provided, this adds theoretical surface elevations based on plane equation: surface.z and bed.z based on surface.z

if \code{out == "none"}, no output file is created but the spatialPointsDataFrame is returned on exit; all other options produce output files but return empty.

\code{outfile} defaults to NULL, meaning that the file name of the output file is automatically created form the first ADCP *.mat file used for input.

fileAppendix is intended to be used together with automatic file names (outfile = NULL) to allow customisation of file names while retaining the convenience of automatically generated file names
}
\author{
Claude Flener \email{claude.flener@utu.fi}
}

