% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graduallyvariedflow_v2.r
\docType{package}
\name{rivr-package}
\alias{rivr-package}
\title{This package is designed as an educational tool for students and
instructors of undergraduate courses in open channel hydraulics.
Functions are provided for computing normal and critical depths,
steady (e.g. backwater curves) and unsteady (flood wave routing)
flow computations for prismatic trapezoidal channels of arbitrary
geometry.}
\description{
This package is designed as an educational tool for students and
instructors of undergraduate courses in open channel hydraulics.
Functions are provided for computing normal and critical depths,
steady (e.g. backwater curves) and unsteady (flood wave routing)
flow computations for prismatic trapezoidal channels of arbitrary
geometry.
}

