% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/traits.R
\name{rivr_traits}
\alias{rivr_trait_length_finite}
\alias{rivr_trait_length_known}
\alias{rivr_traits}
\title{Traits}
\usage{
rivr_trait_length_finite(x, ...)

rivr_trait_length_known(x, ...)
}
\arguments{
\item{x}{Object}

\item{...}{Additional arguments}
}
\description{
Traits
}

