% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transform_iterator.R
\name{transform_iterator}
\alias{transform_iterator}
\title{Transform iterator}
\usage{
transform_iterator(it, fun, ...)
}
\arguments{
\item{it}{An iterator}

\item{fun}{A function}

\item{...}{Additional arguments to \code{fun}}
}
\description{
Apply the function f to the elements of it
}

