% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functional.R
\name{distribution_iterator}
\alias{distribution_iterator}
\title{Distribution iterator}
\usage{
distribution_iterator(f, n = 1, ...)
}
\arguments{
\item{f}{A distribution function}

\item{n}{Number of things to sample. I am so sorry.}
}
\description{
Distribution iterator
}

