% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graduallyvariedflow_v2.r
\docType{package}
\name{rivr-package}
\alias{rivr-package}
\title{Steady and Unsteady Open-Channel Flow Computation}
\description{
This package is designed as an educational tool for students and
instructors of undergraduate courses in open channel hydraulics.
Functions are provided for computing normal and critical depths,
steady (e.g. backwater curves) and unsteady (flood wave routing)
flow computations for prismatic trapezoidal channels. See the vignette
to get started.
}

