% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transform_iterator.R
\name{transform_iterator}
\alias{transform_iterator}
\title{Transform iterator}
\usage{
transform_iterator(it, f, ...)
}
\arguments{
\item{it}{An iterator}

\item{f}{A function}
}
\description{
Apply the function f to the elements of it
}

